+++ src/main/java/com/ilta/solepli/domain/auth/service/JwtTokenProvider.java
@@ -6,7 +6,6 @@ import java.util.Date;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Component;
 
-import io.jsonwebtoken.JwtException;
 import io.jsonwebtoken.Jwts;
 import io.jsonwebtoken.SignatureAlgorithm;
 import io.jsonwebtoken.security.Keys;
@@ -38,25 +37,5 @@ public class JwtTokenProvider {
         .setExpiration(expiry)
         .signWith(key, SignatureAlgorithm.HS256)
         .compact();
-  }
-
-  public String getLoginIdFromToken(String token) {
-    // 토큰에서 사용자 ID 추출
-    return Jwts.parserBuilder()
-        .setSigningKey(key)
-        .build()
-        .parseClaimsJws(token)
-        .getBody()
-        .getSubject();
-  }
-
-  public boolean validateToken(String token) {
-    try {
-      // 토큰 파싱 → 유효한 경우 true
-      Jwts.parserBuilder().setSigningKey(key).build().parseClaimsJws(token);
-      return true;
-    } catch (JwtException e) {
-      return false;
-    }
   }
 }
